Rename to must-colocate-fragment-spreads
